% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cm_to_in.R
\name{cm_to_in}
\alias{cm_to_in}
\title{Calculate value from centimetres to inches}
\usage{
cm_to_in(x)
}
\arguments{
\item{x}{numerical}
}
\value{
numerical
}
\description{
Calculate value from centimetres to inches
}
\examples{
x <- 10
cm_to_in(x)
}
