% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/in_to_cm.R
\name{in_to_cm}
\alias{in_to_cm}
\title{Calculate value from inches to centimetres}
\usage{
in_to_cm(x)
}
\arguments{
\item{x}{numerical}
}
\value{
numerical
}
\description{
Calculate value from inches to centimetres
}
\examples{
x <- 10
in_to_cm(x)
}
